langfuce_test/requirements.txt
langfuse==0.5.0
ollama==0.1.7
python-dotenv==1.0.0
requests==2.31.0 